+++ src/main/java/com/example/Proyect_Beta/repositorio/RepoEmpresa.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepoEmpresa extends JpaRepository<Empresa, Integer>
+public interface RepoEmpresa extends JpaRepository<Empresa, Long>
 {
